+++ database/stakeibc_test.go
@@ -5,9 +5,9 @@ import (
 
 	stakeibctypes "github.com/Stride-Labs/stride/v6/x/stakeibc/types"
 
-	"github.com/forbole/bdjuno/v3/types"
+	"github.com/forbole/bdjuno/v4/types"
 
-	dbtypes "github.com/forbole/bdjuno/v3/database/types"
+	dbtypes "github.com/forbole/bdjuno/v4/database/types"
 )
 
 func (suite *DbTestSuite) TestBigDipperDb_SaveStakeIBCParams() {