+++ modules/gov/handle_genesis.go
@@ -6,9 +6,10 @@ import (
 
 	tmtypes "github.com/tendermint/tendermint/types"
 
-	"github.com/forbole/bdjuno/v3/types"
+	"github.com/forbole/bdjuno/v4/types"
 
-	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
+	gov "github.com/cosmos/cosmos-sdk/x/gov/types"
+	govtypesv1beta1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1beta1"
 	"github.com/rs/zerolog/log"
 )
 
@@ -17,23 +18,23 @@ func (m *Module) HandleGenesis(doc *tmtypes.GenesisDoc, appState map[string]json
 	log.Debug().Str("module", "gov").Msg("parsing genesis")
 
 	// Read the genesis state
-	var genState govtypes.GenesisState
-	err := m.cdc.UnmarshalJSON(appState[govtypes.ModuleName], &genState)
+	var genStatev1beta1 govtypesv1beta1.GenesisState
+	err := m.cdc.UnmarshalJSON(appState[gov.ModuleName], &genStatev1beta1)
 	if err != nil {
 		return fmt.Errorf("error while reading gov genesis data: %s", err)
 	}
 
 	// Save the proposals
-	err = m.saveProposals(genState.Proposals, doc)
+	err = m.saveGenesisProposals(genStatev1beta1.Proposals, doc)
 	if err != nil {
 		return fmt.Errorf("error while storing genesis governance proposals: %s", err)
 	}
 
 	// Save the params
-	err = m.db.SaveGovParams(types.NewGovParams(
-		types.NewVotingParams(genState.VotingParams),
-		types.NewDepositParam(genState.DepositParams),
-		types.NewTallyParams(genState.TallyParams),
+	err = m.db.SaveGenesisGovParams(types.NewGenesisGovParams(
+		types.NewGenesisVotingParams(&genStatev1beta1.VotingParams),
+		types.NewGenesisDepositParam(&genStatev1beta1.DepositParams),
+		types.NewGenesisTallyParams(&genStatev1beta1.TallyParams),
 		doc.InitialHeight,
 	))
 	if err != nil {
@@ -43,8 +44,8 @@ func (m *Module) HandleGenesis(doc *tmtypes.GenesisDoc, appState map[string]json
 	return nil
 }
 
-// saveProposals save proposals from genesis file
-func (m *Module) saveProposals(slice govtypes.Proposals, genDoc *tmtypes.GenesisDoc) error {
+// saveGenesisProposals save proposals from genesis file
+func (m *Module) saveGenesisProposals(slice govtypesv1beta1.Proposals, genDoc *tmtypes.GenesisDoc) error {
 	proposals := make([]types.Proposal, len(slice))
 	tallyResults := make([]types.TallyResult, len(slice))
 	deposits := make([]types.Deposit, len(slice))