+++ modules/stakeibc/module.go
@@ -4,8 +4,8 @@ import (
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/forbole/juno/v4/modules"
 
-	"github.com/forbole/bdjuno/v3/database"
-	stakeibcsource "github.com/forbole/bdjuno/v3/modules/stakeibc/source"
+	"github.com/forbole/bdjuno/v4/database"
+	stakeibcsource "github.com/forbole/bdjuno/v4/modules/stakeibc/source"
 )
 
 var (