+++ modules/stakeibc/source/local/source.go
@@ -7,7 +7,7 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/forbole/juno/v4/node/local"
 
-	stakeibcsource "github.com/forbole/bdjuno/v3/modules/stakeibc/source"
+	stakeibcsource "github.com/forbole/bdjuno/v4/modules/stakeibc/source"
 )
 
 var (